fix(cqrs): virsualization typo
