+++ packages/cqrs/src/commands/index.ts
@@ -43,4 +43,4 @@ export * from './update-record/index.js'
 export * from './update-records/index.js'
 export * from './update-table/index.js'
 export * from './update-view-name/index.js'
-export * from './update-virsualization/index.js'
+export * from './update-visualization/index.js'
