--- packages/cqrs/src/commands/update-virsualization/index.ts
@@ -1,4 +1,0 @@
-export * from './update-virsualization.command.handler.js'
-export * from './update-virsualization.command.input.js'
-export * from './update-virsualization.command.interface.js'
-export * from './update-virsualization.command.js'
