--- packages/cqrs/src/commands/update-virsualization/update-virsualization.command.input.ts
@@ -1,7 +1,0 @@
-import { tableIdSchema, updateVirsualizationSchema } from '@undb/core'
-import { z } from 'zod'
-
-export const updateVirsualizationCommandInput = z.object({
-  tableId: tableIdSchema,
-  virsualization: updateVirsualizationSchema,
-})
