--- packages/cqrs/src/commands/update-virsualization/update-virsualization.command.interface.ts
@@ -1,4 +1,0 @@
-import type { z } from 'zod'
-import type { updateVirsualizationCommandInput } from './update-virsualization.command.input.js'
-
-export type IUpdateVirsualizationCommandInput = z.infer<typeof updateVirsualizationCommandInput>
