+++ packages/cqrs/src/commands/update-visualization/index.ts
@@ -1,0 +1,4 @@
+export * from './update-visualization.command.handler.js'
+export * from './update-visualization.command.input.js'
+export * from './update-visualization.command.interface.js'
+export * from './update-visualization.command.js'
