+++ packages/cqrs/src/commands/update-visualization/update-visualization.command.input.ts
@@ -1,0 +1,7 @@
+import { tableIdSchema, updateVisualizationSchema } from '@undb/core'
+import { z } from 'zod'
+
+export const updateVisualizationCommandInput = z.object({
+  tableId: tableIdSchema,
+  visualization: updateVisualizationSchema,
+})
