+++ packages/cqrs/src/commands/update-visualization/update-visualization.command.interface.ts
@@ -1,0 +1,4 @@
+import type { z } from 'zod'
+import type { updateVisualizationCommandInput } from './update-visualization.command.input.js'
+
+export type IUpdateVisualizationCommandInput = z.infer<typeof updateVisualizationCommandInput>
