+++ packages/cqrs/src/queries/aggregate-number/aggregate-number.query.schema.ts
@@ -1,8 +1,8 @@
-import { tableIdSchema, viewIdSchema, virsualizationIdSchema } from '@undb/core'
+import { tableIdSchema, viewIdSchema, visualizationIdSchema } from '@undb/core'
 import * as z from 'zod'
 
 export const aggregateNumberQuerySchema = z.object({
   tableId: tableIdSchema,
   viewId: viewIdSchema,
-  virsualizationId: virsualizationIdSchema,
+  visualizationId: visualizationIdSchema,
 })
