+++ packages/cqrs/src/queries/aggregate-number/aggregate-number.query.ts
@@ -4,12 +4,12 @@ import type { IAggregateNumberQuery } from './aggregate-number.query.interface.j
 export class AggregateNumberQuery extends Query {
   public readonly tableId: string
   public readonly viewId: string
-  public readonly virsualizationId: string
+  public readonly visualizationId: string
 
   constructor(query: IAggregateNumberQuery) {
     super()
     this.tableId = query.tableId
     this.viewId = query.viewId
-    this.virsualizationId = query.virsualizationId
+    this.visualizationId = query.visualizationId
   }
 }
