+++ packages/cqrs/src/queries/get-chart-data/get-chart-data.query.handler.ts
@@ -1,4 +1,4 @@
-import { ChartVirsualization, type IRecordAggregateRepository, type ITableRepository } from '@undb/core'
+import { ChartVisualization, type IRecordAggregateRepository, type ITableRepository } from '@undb/core'
 import type { IQueryHandler } from '@undb/domain'
 import type { IGetChartDataOutput } from './get-chart-data.query.interface.js'
 import type { GetChartDataQuery } from './get-chart-data.query.js'
@@ -9,11 +9,11 @@ export class GetChartDataQueryHandler implements IQueryHandler<GetChartDataQuery
   async execute(query: GetChartDataQuery): Promise<IGetChartDataOutput> {
     const table = (await this.tableRepo.findOneById(query.tableId)).unwrap()
     const view = table.mustGetView(query.viewId)
-    const virsualization = view.mustGetVirsualization(query.virsualizationId)
+    const visualization = view.mustGetVisualization(query.visualizationId)
 
-    if (!(virsualization instanceof ChartVirsualization)) throw new Error('invalid chart virsualizatio')
+    if (!(visualization instanceof ChartVisualization)) throw new Error('invalid chart virsualizatio')
 
-    const data = await this.repo.chart(query.tableId, virsualization, null)
+    const data = await this.repo.chart(query.tableId, visualization, null)
 
     return { data }
   }
