+++ packages/cqrs/src/queries/get-chart-data/get-chart-data.query.schema.ts
@@ -1,8 +1,8 @@
-import { tableIdSchema, viewIdSchema, virsualizationIdSchema } from '@undb/core'
+import { tableIdSchema, viewIdSchema, visualizationIdSchema } from '@undb/core'
 import * as z from 'zod'
 
 export const getChartDataQuerySchema = z.object({
   tableId: tableIdSchema,
   viewId: viewIdSchema,
-  virsualizationId: virsualizationIdSchema,
+  visualizationId: visualizationIdSchema,
 })
