+++ packages/cqrs/src/queries/get-chart-data/get-chart-data.query.ts
@@ -4,12 +4,12 @@ import type { IGetChartDataQuery } from './get-chart-data.query.interface.js'
 export class GetChartDataQuery extends Query {
   public readonly tableId: string
   public readonly viewId: string
-  public readonly virsualizationId: string
+  public readonly visualizationId: string
 
   constructor(query: IGetChartDataQuery) {
     super()
     this.tableId = query.tableId
     this.viewId = query.viewId
-    this.virsualizationId = query.virsualizationId
+    this.visualizationId = query.visualizationId
   }
 }
